Split and moved some key deserialisation tests in PeerIdTest
